Convert Main from ApplicationAdapter to Game class
